remove implicit unwrapped dictionary command
