+++ SpeedyContactAdd WatchKit App/Base.lproj/Interface.storyboard
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder.WatchKit.Storyboard" version="3.0" toolsVersion="14109" targetRuntime="watchKit" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="AgC-eL-Hgc">
-    <device id="watch38" orientation="portrait">
+<document type="com.apple.InterfaceBuilder.WatchKit.Storyboard" version="3.0" toolsVersion="14460.31" targetRuntime="watchKit" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="AgC-eL-Hgc">
+    <device id="watch44" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
     <dependencies>
         <deployment identifier="watchOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBWatchKitPlugin" version="14031"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBWatchKitPlugin" version="14460.16"/>
     </dependencies>
     <scenes>
         <!--Interface Controller-->
@@ -16,9 +16,9 @@
                     <items>
                         <group width="150" height="20" alignment="left" contentMode="scaleAspectFit" id="QNO-hE-ZFI">
                             <items>
-                                <label width="150" height="20" alignment="left" verticalAlignment="center" textAlignment="center" id="Myc-F5-oVG">
+                                <label width="143" height="20" alignment="center" verticalAlignment="center" text="Enter Number" textAlignment="center" id="Myc-F5-oVG">
                                     <color key="textColor" red="0.93725490199999995" green="0.93725490199999995" blue="0.95686274510000002" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                                    <fontDescription key="font" type="system" pointSize="16"/>
+                                    <fontDescription key="font" type="system" pointSize="18"/>
                                 </label>
                             </items>
                         </group>
@@ -133,12 +133,13 @@
                     </menu>
                     <edgeInsets key="margins" left="0.0" right="0.0" top="0.0" bottom="0.0"/>
                     <connections>
+                        <outlet property="keypadO" destination="dir-4g-4jh" id="zf7-Kh-qOP"/>
                         <outlet property="lblPhoneNumber" destination="Myc-F5-oVG" id="G2g-mG-MBZ"/>
                         <outlet property="lblRecipName" destination="cPN-Iz-YF7" id="UYI-0T-rWI"/>
                     </connections>
                 </controller>
             </objects>
-            <point key="canvasLocation" x="234.51282051282053" y="608.94871794871801"/>
+            <point key="canvasLocation" x="234.30434782608694" y="608.66071428571433"/>
         </scene>
         <!--Static Notification Interface Controller-->
         <scene sceneID="AEw-b0-oYE">
