+++ SpeedyContactAdd WatchKit Extension/Controller/ComplicationController.swift
@@ -60,7 +60,8 @@ import ClockKit
         // Call the handler with the date when you would next like to be given the opportunity to update your complication content
     //    handler(nil);
     }
-    
+
+    // MARK: - Complications for App
      func getPlaceholderTemplate(for complication: CLKComplication, withHandler handler: (CLKComplicationTemplate?) -> Void ) {
